fix rule config handling
